Done and Done
Everything is finished. I might go back and try to do one or two more
things.

I would like a little input on a couple things:

1.  What are a couple of things I could do to clean this up and make it
more organized so that it is easier to read?

2.  Is it possible to change the icon based on the final popup text
input? (This seems tricky since its part of a feature group, and the way
popups work.)